Launch program improved
Now the `launch_asc.ks` file smooth the throttle when reaching the target apoapsis (that made it more precise), also if the target apoapsis is reached while the ship still in the atmosphere, it will keep burning to compensate the effects of drag, if needed.
Also, it now discards fairings automatically, even if they aren't in the right stage.