Protects DOACTION() calls with Career():CANDOACTIONS
This patch wraps all DOACTION() lines with a check for whether
the action is permitted at this stage of career mode.

The boot logic is a guess as that case has not occurred to me yet.

The lib_parts logic does work for me with a VAB that has not yet
been upgraded.