Refactored common code in lib_parts
The library includes a number of functions which traverse the ship
looking for modules by name to perform actions by name, and optionally
return success or failure.  I created an abstract function which does
the needful and updated the existing functions to call that abstract
function.

The abstract function has an additional feature: when supplied with
an optional tag, only parts with that tag will have actions executed.

There are additional functions which extend and retract solar panels.

The code has shrunk by 25% as a result.