Deleting boot script after load
Now that the boot script can be deleted, it can be longer.

No need for a separate installation script, no need to strip the
boot equivalents of the lib_ui functions.

Tracking which files are not found and which fail to copy, plus
performing freespace check before ach copy.  A tiny bit slower but
the results are awesome.